+++ database/migrations/2025_10_23_125851_create_product_mappings_table.php
@@ -13,21 +13,23 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('product_mappings', function (Blueprint $table) {
-            $table->id(); // bigint auto-increment / pk
-            $table->string('distributor_code', 15);
-            $table->string('product_code_dist', 255)->nullable();
-            $table->string('product_name_dist', 255)->nullable();
-            $table->string('product_code_prc', 255)->nullable();
-            $table->timestamps();
-
-            // Foreign Key Constraint
-            $table->foreign('distributor_code')
-                  ->references('distributor_code')
-                  ->on('master_distributors')
-                  ->onUpdate('cascade')
-                  ->onDelete('cascade'); // Hapus mapping jika distributor dihapus
-        });
+        if (!Schema::hasTable('product_mappings')) {
+            Schema::create('product_mappings', function (Blueprint $table) {
+                $table->id(); // bigint auto-increment / pk
+                $table->string('distributor_code', 15);
+                $table->string('product_code_dist', 255)->nullable();
+                $table->string('product_name_dist', 255)->nullable();
+                $table->string('product_code_prc', 255)->nullable();
+                $table->timestamps();
+    
+                // Foreign Key Constraint
+                $table->foreign('distributor_code')
+                      ->references('distributor_code')
+                      ->on('master_distributors')
+                      ->onUpdate('cascade')
+                      ->onDelete('cascade'); // Hapus mapping jika distributor dihapus
+            });
+        }
     }
 
     /**
